requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==1.10.15
fastmcp==0.2.1
python-dotenv==1.0.1
